requirements.txt
Flask
Pillow
requests
sqlite3
tk
